Wrap each tool run in its own process to properly measure memory
Previously all the tool runs were spawned from the same parent process.
This meant that the measured max memory usage was the max memory usage
for any previous tool run, not the one we were trying to measure. By
wrapping the tool runs in their own parent processes we ensure that the
max memory usage is the max memory usage for the command we just ran.