Refactor filter metric counts into separate function
Also fixes bugs with the method that result in the number of counted
reads not being taken into account. Now prints out a warning in the case
that more reads were detected in the out_map than were in the true_map.
It sounds like this case should never happen when testing read
filtering.